% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_covariates.R
\name{add_time_covariates}
\alias{add_time_covariates}
\title{Add some covartiates regarded with time}
\usage{
add_time_covariates(
  x,
  time_col = "time",
  what = c("month", "month_fact", "month_circ", "season", "season_fact", "season_circ")
)
}
\arguments{
\item{x}{A \code{data.frame} with some time (\code{Date}/\code{POSIX}) column.}

\item{time_col}{Name of the time column. It must be of class \code{Date} or 
\code{POSIX}.}

\item{what}{What extra columns you must want to add?}
}
\value{
It returns the same \code{data.frame} with extra columns.
}
\description{
Add some covartiates regarded with time
}
\details{
At the current version, \code{add_time_covariates} allows to include the next 
variables:

\itemize{
 \item \code{month}: Month as \code{numeric}.
 \item \code{month_fact}: Month as \code{factor}.
 \item \code{month_circ}: Month as \code{numeric} but in a cyclic scale.
 \item \code{season}: Season as \code{numeric}.
 \item \code{season_fact}: Season as \code{factor}.
 \item \code{season_circ}: Season as \code{numeric} but in a cyclic scale.
}

If some of the requested variables are already present in \code{x}, you could
indicate a new name by setting a name for that column. For instance, if 
\code{month} column name is already present in \code{x}, you must indicate
\code{what = c(new_month = "month")}.
}
\examples{
exDF <- data.frame(a = 1:10, t = Sys.Date()) 
exDF2 <- add_time_covariates(x = exDF, time_col = "t")
exDF

add_time_covariates(x = exDF2, time_col = "t", what = c(month2 = "month"))
}
