% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gts_data.R
\name{download_gts_data}
\alias{download_gts_data}
\title{Download example data from GitHub repository}
\usage{
download_gts_data(
  files = NULL,
  destdir = ".",
  repo = "LuisLauM/gts-raw-data",
  ref = "main",
  overwrite = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{files}{\code{character} vector with the names of available files. If
\code{NULL}, the list of available files will be showed.}

\item{destdir}{\code{character}. Local directory where the file will be saved.}

\item{repo}{\code{character}. GitHub repository in the form "user/repo".}

\item{ref}{\code{character}. Branch, tag, or commit (default: "main").}

\item{overwrite}{\code{logical}. Whether to overwrite existing files.}

\item{quiet}{If \code{TRUE}, suppress downloading status messages (if any), 
and the progress bar.}
}
\value{
A \code{list} with paths to the downloaded files.
}
\description{
This function downloads files stored in the package GitHub repository
using the raw GitHub URL (raw.githubusercontent.com).
}
\examples{
download_gts_data()

download_gts_data(files = c("bathymetry.nc", "shelfbreak.nc"))
}
