% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{gts-reexports}
\alias{gts-reexports}
\alias{melt}
\title{Re-exported functions}
\usage{
melt(data, ..., na.rm = FALSE, value.name = "value")
}
\description{
These functions are imported from other packages and re-exported for
convenience.
}
\keyword{internal}
