% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate.gts}
\alias{integrate.gts}
\title{Vertical integration of a gridded time series.}
\usage{
\method{integrate}{gts}(f, lower, upper = 0, by = NULL, subdivisions = NULL, ...)
}
\arguments{
\item{f}{An object of class gts.}

\item{lower}{Lower limit for vertical integration.}

\item{upper}{Upper limit for vertical integration.}

\item{by}{Width of the integration interval, 1 by default.}

\item{subdivisions}{Number of subdivisions to be used}

\item{...}{Additional arguments, currently unused.}
}
\value{
The original object after vertical integration
}
\description{
Vertical integration of a gridded time series.
}
\examples{
integrate()
}
