% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{is_over}
\alias{is_over}
\title{Find if points are over a layer (Spatial Polygon)}
\usage{
is_over(lon, lat, layer)
}
\arguments{
\item{lon}{Vector of longitudes.}

\item{lat}{Vector of latitudes.}

\item{layer}{A spatialPolygons object}
}
\value{
A data.frame with TRUE/FALSE.
}
\description{
Find if points are over a layer (Spatial Polygon)
}
