% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-kali.R
\name{map.axes2}
\alias{map.axes2}
\title{Add geographical axes to a map}
\usage{
map.axes2(sides = c(1, 2), cex.axis = 0.75, line = -0.4)
}
\arguments{
\item{sides}{Side of the axis, 1 and 2 by default. See \code{graphics::axis}.}

\item{cex.axis}{Character expansion factor for axis labels.}

\item{line}{the number of lines into the margin at which the axis line
    will be drawn, if not \code{NA}.}
}
\description{
Add geographical axes to a map
}
