% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-kali.R
\name{map_details}
\alias{map_details}
\title{Add details to a map}
\usage{
map_details(
  primeMeridian = "center",
  hires = FALSE,
  col = "darkolivegreen4",
  interior = FALSE,
  axes = FALSE,
  border = FALSE,
  boundaries.col = "black",
  grid = FALSE,
  grid.col = "white",
  grid.lwd = 0.5,
  cex.axis = 0.75,
  fill = TRUE,
  boundary = TRUE,
  water = NULL,
  land = TRUE,
  countries = FALSE,
  nx = NULL,
  ny = nx,
  labels = TRUE,
  ...
)
}
\arguments{
\item{primeMeridian}{Prime Meridian used in the map, possible values are "center" and "left", the former is the default.}

\item{hires}{Boolean, use high resolution coastline?}

\item{col}{Color table to use for image, by default rev(rainbow(nlevel/10, start = 0/6, end = 4/6))}

\item{interior}{
    If \code{FALSE}, interior segments are not drawn.
    An interior segment is a line segment of the map that bounds two
    of the polygons to be drawn.
    This argument is ignored if \code{fill} is \code{TRUE}.
  }

\item{axes}{Boolean, add axes to the plot? Default is TRUE.}

\item{border}{This only works if x and y are matrices -- if NA the quadralaterals will 
have a border color that is 
the same as the interior color. Otherwise this specifies the color to use.}

\item{boundaries.col}{Color of the country boundaries.}

\item{grid}{Boolean, add grid lines over the map?}

\item{grid.col}{Color of the grid.}

\item{grid.lwd}{Width of the grid lines.}

\item{fill}{
    logical flag that says whether to draw lines or fill areas.
    If \code{FALSE}, the lines bounding each region will be drawn (but
    only once, for interior lines).
    If \code{TRUE}, each region will be filled using colors from the
    \code{col =} argument, and bounding lines are drawn by default using \code{par("fg")}.
    To hide the bounding lines, use \code{border=NA} (see \code{...}).
  }

\item{boundary}{
    If \code{FALSE}, boundary segments are not drawn.
    A boundary segment is a line segment of the map that bounds only
    one of the polygons to be drawn.
    This argument is ignored if \code{fill} is \code{TRUE}.
  }

\item{water}{Color of the water masses.}

\item{land}{Boolean, add a polygon over the land? If FALSE, only coastline is added.}

\item{countries}{Boolean, do add country boundaries? Default is FALSE.}

\item{nx, ny}{number of cells of the grid in x and y direction.  When
    \code{NULL}, as per default, the grid aligns with the tick marks on
    the corresponding \emph{default} axis (i.e., tickmarks as computed by
    \code{\link[graphics]{axTicks}}).  When \code{\link{NA}}, no grid lines are
    drawn in the corresponding direction.}

\item{labels}{Boolean, add LATITUDE and LONGITUDE labels?}

\item{...}{}
}
\description{
Add details to a map
}
