% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_gts}
\alias{merge_gts}
\title{Merge a data.frame and a GTS object.}
\usage{
merge_gts(x, y, dt = "month", dimnames = NULL, ...)
}
\arguments{
\item{x}{A data frame with dates and coordinates.}

\item{y}{A GTS object.}

\item{dt}{Time step when only 1 time value is used (the default is "month").}

\item{dimnames}{The names of the dimensions in the data.frame.}

\item{...}{Additional arguments.}
}
\value{
A data.frame with the corresponding values of the variable from the GTS object.
}
\description{
Merge a data.frame and a GTS object.
}
