% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{read_grid}
\alias{read_grid}
\title{Read the grid from a file}
\usage{
read_grid(filename, varid = NULL, mask = NULL, create_mask = FALSE, ...)
}
\arguments{
\item{filename}{The filename of the netCDF file.}

\item{varid}{Variable used to create the mask, see details.}

\item{mask}{Variable with an actual mask for the grid, must be of dimension 2.}

\item{create_mask}{Should we create a new mask using the same algorithm as \link{make_grid}?}

\item{...}{Additional arguments passed to the \link{make_grid} function, mainly 'n', 'thr' and 'hires'.}
}
\value{
Return the grid associated to the variables in the file.
}
\description{
Read the grid from a file
}
\examples{
read_grid()
}
