% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts-class.R
\name{read_gts}
\alias{read_gts}
\title{Read a netCDF variable into a Gridded Time-Series.}
\usage{
read_gts(filename, varid = NULL, control = list(), ...)
}
\arguments{
\item{filename}{The filename of the netCDF file.}

\item{varid}{String with  the name of the variable in the file to read the
data from. If left unspecified, the function will determine if there is only
one variable in the file and, if so, read from that. If left unspecified and
there are multiple variables in the file, an error is generated.}

\item{control}{A list with additional options, particularly the time units and origin when not specified in the file.}

\item{...}{Additional arguments.}
}
\value{
A gts object
}
\description{
Read a netCDF variable into a Gridded Time-Series.
}
