% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{read_static}
\alias{read_static}
\title{Read an static variable from a netCDF file}
\usage{
read_static(filename, varid = NULL, create_mask = FALSE, ...)
}
\arguments{
\item{filename}{The filename of the netCDF file.}

\item{varid}{Variable to read static variable (time-invariant).}

\item{create_mask}{Should we create a new mask using the same algorithm as \link{make_grid}?}

\item{...}{Additional arguments, currently unused.}
}
\value{
An object of class 'static'
}
\description{
Read an static variable from a netCDF file
}
