% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded-inspect.R
\name{setnames_gts}
\alias{setnames_gts}
\title{Replace variable identifier and long name}
\usage{
setnames_gts(x, value)
}
\arguments{
\item{x}{A `gts` or `static` object.}

\item{value}{A character vector containing the variable identifier and long
name.}
}
\value{
The modified object.
}
\description{
Internal helper used by the `names<-` methods for gridded objects.
}
\keyword{internal}
