% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded-inspect.R
\name{setunits_gts}
\alias{setunits_gts}
\title{Replace data-variable units}
\usage{
setunits_gts(x, value)
}
\arguments{
\item{x}{A `gts` or `static` object.}

\item{value}{A character scalar giving the units of the data variable.}
}
\value{
The modified object.
}
\description{
Internal helper used by the `units<-` methods for gridded objects.
}
\keyword{internal}
