% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{stl_array2}
\alias{stl_array2}
\title{Seasonal Decomposition of Gridded Time Series}
\usage{
stl_array2(x, frequency = 12, s.window = "periodic", ...)
}
\arguments{
\item{x}{An array with a time dimension.}

\item{frequency}{The frequency in time-steps per year, default to 12.}

\item{s.window}{Only "periodic" is supported now.}

\item{...}{Additional arguments controling the methods.}
}
\value{
A list, with the trend, seasonal and residual components.
}
\description{
Seasonal Decomposition of Gridded Time Series
}
