% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.gts}
\alias{subset.gts}
\alias{subset.grid}
\alias{subset.static}
\title{Subsetting a spatial time series object}
\usage{
\method{subset}{gts}(x, longitude = NULL, latitude = NULL, grid = NULL, expand = 0, ...)

\method{subset}{grid}(
  x,
  longitude = NULL,
  latitude = NULL,
  index.return = FALSE,
  grid = NULL,
  expand = 0,
  ...
)

\method{subset}{static}(x, longitude = NULL, latitude = NULL, grid = NULL, expand = 0, ...)
}
\arguments{
\item{x}{the object to be subsetted.}

\item{longitude}{The range of longitude to subset the grid.}

\item{latitude}{The range of latitude to subset the grid.}

\item{grid}{A grid object, or a 'gts' object.}

\item{expand}{Number of units to expand the grid, generally degrees.
It can be one number or two (longitude, latitude).}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A 'gts' object after subsetting.
}
\description{
Subsetting a spatial time series object
}
