% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{ts_accuracy}
\alias{ts_accuracy}
\title{Get in-sample accuracy of all models}
\usage{
ts_accuracy(.ts)
}
\arguments{
\item{.ts}{A `gts` mable}
}
\value{
A tibble
}
\description{
Get in-sample accuracy of all models from a `gts` mable
}
\examples{
your_mbl <- your_ts \%>\%
  ts_prep() \%>\%
  ts_model(forecast::auto.arima) \%>\%
  ts_model(forecast::ets, with_log = TRUE)

your_mbl \%>\%
  ts_accuracy()
}
