% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{ts_forecast}
\alias{ts_forecast}
\title{Create forecasts}
\usage{
ts_forecast(.ts, h = NULL)
}
\arguments{
\item{.ts}{A `gts` mable}

\item{h}{The number of periods to forecast. If NULL, forecasts the frequency of the input time-series.}
}
\value{
A tibble
}
\description{
Create forecasts from a `gts` mable.
}
\examples{
your_ts \%>\%
  ts_prep() \%>\%
  ts_model(forecast::auto.arima) \%>\%
  ts_model(forecast::ets, with_log = TRUE)

your_mbl \%>\%
  ts_forecast(h = 12)
}
