% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ts_model}
\alias{ts_model}
\title{Train a model}
\usage{
ts_model(.data, ...)
}
\arguments{
\item{...}{forecast functions passed}

\item{.ts}{A `gts`` prepped time-series}
}
\value{
A gts mable
}
\description{
Train a model exported from the \code{forecast} package
}
\examples{
your_ts \%>\%
  ts_prep() \%>\%
  ts_model(forecast::auto.arima) \%>\%
}
