% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{ts_prep}
\alias{ts_prep}
\title{Convert tsibble to forecast ready object}
\usage{
ts_prep(.ts)
}
\arguments{
\item{.ts}{A tsibble}
}
\value{
A nested data frame
}
\description{
Convert tsibble to forecast ready object
}
\examples{
your_ts \%>\%
  ts_prep()
}
