% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{ts_split}
\alias{ts_split}
\title{Split data}
\usage{
ts_split(.data, times = 8)
}
\arguments{
\item{.data}{A nested tibble}

\item{times}{Number of equal sized groups to make}
}
\description{
Split data into a list of equal sized groups
}
