% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_fable.R
\name{ts_unnest_fable}
\alias{ts_unnest_fable}
\title{Unnest a fable column}
\usage{
ts_unnest_fable(.data, col)
}
\arguments{
\item{.data}{A nested tibble with a fable column}

\item{col}{The fable column to unnest}

\item{keep}{Other columns to keep}
}
\description{
Unnest a fable column
}
