% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Calculate the area of every square of a grid}
\usage{
area(x, ...)
}
\arguments{
\item{x}{A grid, linear o curvilinear.}

\item{...}{Additional parameters for grid computation.}
}
\value{
A matrix, with the area of each square of the grid.
}
\description{
Calculate the area of every square of a grid
}
