% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts-auxiliar.R
\name{compact_gam}
\alias{compact_gam}
\title{Reduce the object of a GAM object by eliminating caller environment references}
\usage{
compact_gam(object, ...)
}
\arguments{
\item{object}{A GAM or GLM object}

\item{...}{Additional arguments}
}
\description{
Reduce the object of a GAM object by eliminating caller environment references
}
