% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\alias{fill.default}
\alias{fill.grid}
\alias{fill.gts}
\title{Fill missing values in an array}
\usage{
fill(x, ...)

\method{fill}{default}(x, method, control = list(), ...)

\method{fill}{grid}(x, method, control = list(), ...)

\method{fill}{gts}(x, method, control = list(), ...)
}
\arguments{
\item{x}{A matrix or a supported array (e.g. a gts::gts object).}

\item{...}{Additional arguments, currently ignored.}

\item{method}{The interpolation method used, currently only 'akima' available.}

\item{control}{Additional control argument passed to the interpolation method.}
}
\value{
The array without missing values.
}
\description{
Fill missing values in an array
}
