% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.gts.R, R/gts-class.R
\name{as.gts}
\alias{as.gts}
\alias{gts}
\alias{gts.ncdf4}
\alias{gts.array}
\title{Create a 'gts' object.}
\usage{
as.gts(x, ...)

gts(x, ...)

\method{gts}{ncdf4}(x, varid = NULL, climatology = FALSE, control = list(), ...)

\method{gts}{array}(
  x,
  varid = NULL,
  grid,
  data = NA,
  start = 1,
  end = numeric(),
  frequency = 1,
  deltat = 1,
  climatology = FALSE,
  long_name = NULL,
  unit = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{x}{An array-like object to be a gridded time-series.}

\item{...}{Additional arguments to be passed to different methods.}

\item{varid}{Name of the variable (e.g. from a ncdf file).}

\item{control}{A list with control arguments for the creation of a gts object. See details.}
}
\value{
A 'gts' object.
}
\description{
Create a 'gts' object.
}
