% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts-class.R
\name{gts}
\alias{gts}
\alias{gts.ncdf4}
\title{Create a 'gts' object.}
\usage{
gts(x, ...)

\method{gts}{ncdf4}(x, varid = NULL, control = list(), ...)
}
\arguments{
\item{x}{An array-like object to be a gridded time-series.}

\item{...}{Additional arguments to be passed to different methods.}

\item{varid}{Name of the variable (e.g. from a ncdf file).}

\item{control}{A list with control arguments for the creation of a gts object. See details.}
}
\value{
A 'gts' object.
}
\description{
Create a 'gts' object.
}
