% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolation of bivariate data}
\usage{
interpolate(x, y, z, ...)
}
\arguments{
\item{x}{x input}

\item{y}{y input}

\item{z}{z input}

\item{...}{Additional arguments}

\item{xout}{x output}

\item{yout}{y output}

\item{method}{The method}

\item{extrap}{Do you want to extrapolate outside the convex-hull of your data?}

\item{control}{Control arguments}
}
\value{
A list
}
\description{
Interpolation of bivariate data
}
\examples{
data(volcano)
x = seq(from=0, to=1, length.out = nrow(volcano))
y = seq(from=0, to=1, length.out = ncol(volcano))
xout = seq(from=0, to=1, length.out = 3*nrow(volcano))
yout = seq(from=0, to=1, length.out = 3*ncol(volcano))
zout = interpolate(x, y, volcano, xout, yout)
}
