% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts-class.R
\name{read_gts}
\alias{read_gts}
\title{Read a netCDF variable into a Gridded Time-Series.}
\usage{
read_gts(filename, varid, control = list(), ...)
}
\arguments{
\item{filename}{The filename of the netCDF file.}

\item{varid}{Name of the variable in the file.}

\item{control}{Additional options, currently not used.}

\item{...}{Additional arguments.}
}
\value{
A gts object
}
\description{
Read a netCDF variable into a Gridded Time-Series.
}
