% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{regrid}
\alias{regrid}
\title{Regridding}
\usage{
regrid(object, grid, ...)
}
\arguments{
\item{object}{The object}

\item{grid}{The grid}

\item{...}{Additional arguments}
}
\value{
A regrided object
}
\description{
Regridding
}
\examples{
regrid(object, grid)
}
