% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{resolution}
\alias{resolution}
\title{Compute the spatial resolution of a grid}
\usage{
resolution(x, ...)
}
\arguments{
\item{x}{A gridded object, gts, static or grid.}

\item{...}{Additional parameters for computation.}
}
\value{
The spatial resolution.
}
\description{
Compute the spatial resolution of a grid
}
