% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{transform_coordinates}
\alias{transform_coordinates}
\title{Transform coordinates between Coordinate Reference Systems (CRS)}
\usage{
transform_coordinates(object, crs_old, crs_new = "EPSG:4326")
}
\arguments{
\item{object}{A two-column matrix with 'x' (~longitude) and 'y' (~latitude) coordinates as columns.}

\item{crs_old}{Original CRS of the coordinates.}

\item{crs_new}{New CRS of the output.}
}
\value{
A data.frame with 'x' and 'y' coordinates in the new CRS.
}
\description{
Transform coordinates between Coordinate Reference Systems (CRS)
}
