% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{ts_accuracy}
\alias{ts_accuracy}
\title{Get in-sample accuracy of all models}
\usage{
ts_accuracy(.data, ...)
}
\arguments{
\item{.ts}{A `gts` mable}
}
\value{
A data.table
}
\description{
Get in-sample accuracy of all models from a `gts` mable
}
\examples{
your_mbl <- your_df \%>\%
  ts_prep(key = c(level_1, level_2), index = index, target = revenue)
  ts_model(ets(.ts),
           ets(log(.ts + 1)))

your_mbl \%>\%
  ts_accuracy()
}
