% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{ts_forecast}
\alias{ts_forecast}
\title{Create forecasts}
\usage{
ts_forecast(.data, h = 12)
}
\arguments{
\item{h}{The number of periods to forecast. Defaults to 12}

\item{.ts}{A `gts` mable}
}
\value{
A tibble
}
\description{
Create forecasts from a `gts` mable.
}
\examples{
your_mbl <- your_df \%>\%
  ts_prep(key = c(level_1, level_2), index = index, target = revenue)
  ts_model(ets(.ts),
           ets(log(.ts + 1)))

your_mbl \%>\%
  ts_forecast(h = 12)
}
