% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ts_model}
\alias{ts_model}
\title{Train a model}
\usage{
ts_model(.ts, .fn, with_log = FALSE, ...)
}
\arguments{
\item{.ts}{A `gts`` prepped time-series}

\item{.fn}{A function exported by the \code{forecast} pacakge}

\item{with_log}{TRUE/FALSE whether a log transform should be applied}

\item{...}{Additional parameters passed to the \code{forecast} function}
}
\value{
A gts mable
}
\description{
Train a model exported from the \code{forecast} package
}
\examples{
your_ts \%>\%
  ts_prep() \%>\%
  ts_model(forecast::auto.arima) \%>\%
  ts_model(forecast::ets, with_log = TRUE)
}
