% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{ts_prep}
\alias{ts_prep}
\title{Convert tsibble to forecast ready object}
\usage{
ts_prep(.data, key = NULL, index = index, target = NULL)
}
\arguments{
\item{.data}{A data.frame or data.table}
}
\description{
Convert tsibble to forecast ready object
}
\examples{
your_df \%>\%
  ts_prep(key = c(level_1, level_2), index = index, target = revenue)
}
