% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{ts_prep}
\alias{ts_prep}
\title{Convert tsibble to forecast ready object}
\usage{
ts_prep(.data, key = NULL, index = index, target = NULL, frequency = 12)
}
\arguments{
\item{.data}{A data.frame or data.table}
}
\value{
A nested data frame
}
\description{
Convert tsibble to forecast ready object
}
\examples{
your_ts \%>\%
  ts_prep()
}
